requirements.txt
einops==0.4.1
numpy==1.24.4
opencv-python==3.4.10.37
Pillow==10.0.1
PyYAML==6.0.2
scikit-learn==1.3.2
scipy==1.5.4
tabulate==0.8.10
tensorboardX==2.2
torch==1.13.0
torchvision==0.14.0
open3d==0.17.0
typing-extensions==4.10.0
timm==0.5.4
easydict==1.11
protobuf==3.20.3
anomalib==0.7.0
